(PIE-391) Add method to retreive custom parameters via classifier api
For now it seems that storing the job_id and and other parameters as a
custom parameter that can be retreived by the classifier api is the best
option. Users will be instruted to create these parameters for each use
case in the relevant module documentation and use this method to
retrieve them.