(MAINT) Turn this repo back into a module
We wanted to make this repo a ruby gem repository only. Unfortunately
that turns out to be a problem for Bolt. So this change brings back the
tasks that were moved out of the repo not that long ago.

It also updates some of the task metadata to ensure it will work
properly in the new structure.