(MAINT) Fix pagination.
Previously only the offset would be included if both a limit and offset
were specified. This adds a few separate method including one to create
a param string from a hash of params, one the create a hash of limit and
offset params, and one to combine the two into a convinience method.