(PIE-408) Release Automation
An informal poll of the team found that we'd prefer a simple push to
forge when a version tag is pushed to github. This means that we can
choose to create a release branch or not, but all this does is know how
to push a module to the forge when a tag is specified.