(PIE-465) Prevent double uri queries
Previously we hard coded the ? character to the beginning of our
pagination params to signify the beginning of a uri query. This caused a
bug where a second ? could be added to the uri if another query was
passed as part of the uri. Now the uri query is completely deconstructed
and remade each time the make_params method is called.