(PIE-475) Allow Events and Orchestrator classes to accept auth token.
Modify the Events and Orchestrator classes to accept an auth token in
order to prevent a sensitive username and password from being stored in
a config file. In this scenario the user provides their own long lived
auth token rather than having the PEHTTP class generate one. Users are
still able to use a username and password if desired.