(PIE-654) Implement Lock File
The coordinator script should wake up and be able to tell if the
previous run has completed. It should then write it's own lock file
to prevent a subsequent concurrent run.

The lock file should be fault tolerant and not leave an orphan lock
file if the script encounters an error.

It should also write enough information to the file so that the script
is not confused if the operating system reuses the PID in the case
that an orphan PID file is left behind for some reason.