(PIE-761) Implement Search for Processors
This change adds a class that knows how to search a specified directory
at only the first level looking for files, but not search recursively
through directories. Those directories may contain support library files
for the processors, not processors.

This class can then be created from a file path and it knows how to
invoke a processor and capture its output streams and exit code.

That data can then be accessed for logging purposes.