(PIE-525) HSBC Defect - First Run Protection
When the module is first started it should not send every event in the
api all at once. It should realize this is the first time it's running,
note the number of available events in the index file and then shut
down.

On the next run it should then only send the new events since that first
run.