(PIE-775) Remove Unnecessary Files
Over time this module has accumulated a lot of files it doesn't need.
Some files are from when this was going to be a gem, some others are
seemingly indecision between using shell scripts vs Bolt tasks, and
some others are artifacts from the fact that this module seems to have
been cloned from the original CMDB integration module.