(MAINT) Document Sensitive param in README
One customer has already noted, with some frustration, that they were
initially unable to get this module working, because they didn't know
that the `pe_password` parameter could not be assigned via the console.

This change ensures that future users will be warned and given a
workaround earlier in the implementation process.