(PIE-1034) Ensure lockfile only deleted when appropriate
The collect_api_events.rb script will check if another instance of the
script is running and exit if so. The problem arises when the exit is
caught so we can create the appropriate log messages and an ensure block
deletes the lockfile. In scenarios where the cron calls start to stack
on each other this ensure clause can delete lockfiles created by other
running instances of collect_api_events.rb.

To solve this we introduce a check before deleting the lockfile to
ensure that the current pid matches the pid in the lockfile. There is a
force parameter to override this check which is used by the
validate_command in the case that the pid referenced in the lockfile no
longer exists.